+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Change Log
 
+# 0.3.2
+
+- Updated build actions to format the results dropping all escape sequences and to return error results also.
+- Added testing framework and tests for all actions
+- Renamed inspect python file to fix issue with nose testing
+  Contributed by Bradley Bishop (Encore Technologies)
+
 # 0.3.0
 
 - Updated action `runner_type` from `run-python` to `python-script`
@@ -10,4 +17,4 @@
 
 # 0.1.0
 
-- First release 
+- First release